requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.7
alembic==1.12.1

# Redis & Caching
redis[hiredis]>=4.0.1
aiofiles==23.2.0

# Background Tasks
celery[redis]>=4.0.1

# Web Scraping
aiohttp==3.8.6
beautifulsoup4==4.12.2
selenium==4.15.2
undetected-chromedriver==3.5.4
playwright==1.40.0
cloudscraper==1.2.71
requests==2.31.0

# Data Processing
pandas==2.1.3
numpy==1.24.3
pydantic[email]==2.4.2
pydantic-settings==2.0.3
thefuzz==0.20.0

# Machine Learning
scikit-learn==1.3.2
joblib==1.3.2
statsmodels==0.14.0

# Visualization
matplotlib==3.8.1
seaborn==0.13.0
plotly==5.17.0

# Monitoring & Logging
prometheus-client==0.19.0
psutil==5.9.6

# Configuration
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-cov==4.1.0

# Development
black==23.10.1
isort==5.12.0
flake8==6.1.0
mypy==1.7.0

# Production
supervisor==4.2.5